Add clear and new grid buttons
Added a 'clear' button which resets the entire grid to a white
background. Added a 'new grid' button which allows the user to input
the size of the grid they want and creates an X-by-X grid based on the
number input by the user. The new grid fits in the same 640px by 640px
container regardless of size.